app/requirements.txt
Django==2.2.10
djangorestframework==3.11.1
django-debug-toolbar==2.2
psycopg2-binary==2.8.5

pytest==6.0.1
pytest-django==3.9.0